refactor: 토큰 재발급 로직을 Auth로 통합
