+++ android/app/src/main/java/com/yagubogu/data/datasource/auth/AuthRemoteDataSource.kt
@@ -2,7 +2,9 @@ package com.yagubogu.data.datasource.auth
 
 import com.yagubogu.data.dto.request.auth.LoginRequest
 import com.yagubogu.data.dto.request.auth.LogoutRequest
+import com.yagubogu.data.dto.request.token.TokenRequest
 import com.yagubogu.data.dto.response.auth.LoginResponse
+import com.yagubogu.data.dto.response.token.TokenResponse
 import com.yagubogu.data.service.AuthApiService
 import com.yagubogu.data.util.safeApiCall
 import javax.inject.Inject
@@ -10,6 +12,12 @@ import javax.inject.Inject
 class AuthRemoteDataSource @Inject constructor(
     private val authApiService: AuthApiService,
 ) : AuthDataSource {
+    override suspend fun refreshToken(refreshToken: String): Result<TokenResponse> =
+        safeApiCall<TokenResponse> {
+            val tokenRequest = TokenRequest(refreshToken)
+            authApiService.postRefresh(tokenRequest)
+        }
+
     override suspend fun login(idToken: String): Result<LoginResponse> =
         safeApiCall<LoginResponse> {
             val loginRequest = LoginRequest(idToken)
