--- android/app/src/main/java/com/yagubogu/data/datasource/token/TokenDataSource.kt
@@ -1,7 +1,0 @@
-package com.yagubogu.data.datasource.token
-
-import com.yagubogu.data.dto.response.token.TokenResponse
-
-interface TokenDataSource {
-    suspend fun refreshTokens(refreshToken: String): Result<TokenResponse>
-}
