--- android/app/src/main/java/com/yagubogu/data/datasource/token/TokenRemoteDataSource.kt
@@ -1,17 +1,0 @@
-package com.yagubogu.data.datasource.token
-
-import com.yagubogu.data.dto.request.token.TokenRequest
-import com.yagubogu.data.dto.response.token.TokenResponse
-import com.yagubogu.data.service.TokenApiService
-import com.yagubogu.data.util.safeApiCall
-import javax.inject.Inject
-
-class TokenRemoteDataSource @Inject constructor(
-    private val tokenApiService: TokenApiService,
-) : TokenDataSource {
-    override suspend fun refreshTokens(refreshToken: String): Result<TokenResponse> =
-        safeApiCall<TokenResponse> {
-            val tokenRequest = TokenRequest(refreshToken)
-            tokenApiService.postRefresh(tokenRequest)
-        }
-}
