+++ android/app/src/main/java/com/yagubogu/data/repository/auth/AuthDefaultRepository.kt
@@ -10,6 +10,16 @@ class AuthDefaultRepository @Inject constructor(
     private val authDataSource: AuthDataSource,
     private val tokenManager: TokenManager,
 ) : AuthRepository {
+    override suspend fun refreshToken(): Result<Unit> {
+        val refreshToken: String =
+            tokenManager.getRefreshToken()
+                ?: return Result.failure(Exception(ERROR_NO_REFRESH_TOKEN))
+
+        return authDataSource.refreshToken(refreshToken).map { (accessToken, refreshToken) ->
+            tokenManager.saveTokens(accessToken, refreshToken)
+        }
+    }
+
     override suspend fun login(idToken: String): Result<LoginResultResponse> =
         authDataSource.login(idToken).map { loginResponse: LoginResponse ->
             tokenManager.saveTokens(loginResponse.accessToken, loginResponse.refreshToken)
