+++ android/app/src/main/java/com/yagubogu/data/repository/auth/AuthRepository.kt
@@ -3,6 +3,8 @@ package com.yagubogu.data.repository.auth
 import com.yagubogu.data.dto.response.auth.LoginResultResponse
 
 interface AuthRepository {
+    suspend fun refreshToken(): Result<Unit>
+
     suspend fun login(idToken: String): Result<LoginResultResponse>
 
     suspend fun logout(): Result<Unit>
