--- android/app/src/main/java/com/yagubogu/data/repository/token/TokenRepository.kt
@@ -1,5 +1,0 @@
-package com.yagubogu.data.repository.token
-
-interface TokenRepository {
-    suspend fun refreshTokens(): Result<Unit>
-}
